feat: implement BarList and UserList components
Add BarList component to manage and display bars with 
functionality to add new bars. Refactor AdminRoute to 
use BarList and UserList components for better code 
organization and maintainability. Introduce user list 
styling for joined status indication.
