+++ src/App.css
@@ -1,7 +1,12 @@
 @import "./components/nav.css" layer(organism);
 @import "./components/header.css" layer(organism);
+
+@import "./routes/unregistered.index.css" layer(page);
+
+@import "./routes/admin.css" layer(page);
+
+@import "./components/barList.css" layer(organism);
 @import "./components/barCard.css" layer(organism);
 @import "./components/addBar.css" layer(organism);
 
-@import "./routes/unregistered.index.css" layer(page);
-@import "./routes/admin.css" layer(page);
+@import "./components/userList.css" layer(organism);
